Fix [adhocracy.lib.instance.discriminator] No such instance: ... bug
We strip the dot form the subdomain in the next line.
We don't need to to do it with slicing, which ends up in
a slice [:-1] instead of [:0] for the main domain and tues
tries to use the main domain name (everything but the last char)
like a instance name).