integrate piwik code again (does not work with diazo atm)
This integrates the piwik code again if a piwik site
is given in the ini file.
It does not work with diazo atm. The piwik code is wrapped
into a <div id="piwik"> so it can be refered to with
diazo rules.
(Maybe better integrate the code into the themes index.html
cause content fechted from wordpress might not be tracked
otherwise.)