Disable Simile's history feature.
It does not work in a local setup and theming.
Simile tries to get "dummy" pages to add entries
to the browser history. These use special urls
postfixed with ',0', ',1', ...
A front end apache seems to strip them so our
static app deliveres /static/__history__.html
If this does not work (theming, 404) we return
a full page with resource links, the browser
refetches these resources and we end up with
a huge delay in the browser.