Fix routing for page history using routes {.format} extension
This makes /page/<id>/history.overlay work.
It also reduces the number of rules from 4 (with and without <...>.{format}
to two with <...>{.format} cause {.format} is optional.

Change the signature of PageController.history() cause routes pathes
format=None to the controller, not format='html', which is the default
for all views using the old .{format}. Doesn't matter for the history.