(Manual) uniqe relation between proposals and page variants
This manually maintains a relation between a proposal and page
*variants* by saving the related variant names on selections. Also
asserting that variants are not related to more than 1 proposal
(except model.Text.HEAD, which is related to all Proposals that select
a paper). This does not use sql relations or constrains, but
PickleType to save a list and is ment as a temporary solution until we
have a new content type <-> variant <-> content schema.

* Add `variants` to the selection schema and preset it to [Text.HEAD].
* Save the variant names on the selection when a variant is added out of
  a proposal. Assert that a variant is not added to two selections.
* Add Selection.by_variant() to find the selection for a variant.